rename loan types to loan options
