+++ src/_layouts/sidebar_nav.html
@@ -3,7 +3,7 @@
     ('../', 'index', 'Owning a Home'),
     ('../rate-checker/', 'rate-checker', 'Rate Checker'),
     ('../loan-comparison/', 'loan-comparison', 'Loan Comparison'),
-    ('../loan-types/', 'loan-types', 'Loan types')
+    ('../loan-options/', 'loan-options', 'Loan Options')
 ] -%}
 {% set active_page = active_page|default('index') -%}
