+++ src/loan-options/index.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
-{% set active_page = "loan-types" %}
+{% set active_page = "loan-options" %}
 
-{% block title %}Owning a Home: Loan Types{% endblock title %}
+{% block title %}Owning a Home: Loan Options{% endblock title %}
 
 {% block content %}
