+++ src/static/js/modules/loan-options.js
